+++ src/main/java/com/schoolmanagement/repository/IFacultyRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface IFacultyRepository extends CrudRepository<Faculty, Integer> {
+public interface IFacultyRepository extends CrudRepository<Faculty, Long> {
     Optional<Faculty> findByName(String name);
 }
